REFACTOR S04_03 refactoring the use of double values to provide consistency since double should not be used for money counting, moved to cents model, so now all cash is represented as int and when display is required an utility method is used to present the form of .2f decimal the actual amount in cents. e.g. 100 in system when shown to user will be 1$ and 678 will be shown to user as 6.78$

+++ src/main/java/tdd/vendingMachine/domain/ShelfItem.java
@@ -15,8 +15,8 @@ public interface ShelfItem {
     String provideType();
 
     /**
-     * The value of the item on the shelve
-     * @return double
+     * The value in cents of the item on the shelve
+     * @return int
      */
-    double provideValue();
+    int provideValue();
 }
